Move state transition open sibling function to separate file
Better separation of the panel and all the imports for the panel and the open sibling function which does
not need all the logic of the panel. This avoids bringing in the imports of the StateTransition panel into other panels and since this
logic is not specific to anything else in the panel I find it can be nicely separated.
