+++ packages/studio-base/src/panels/diagnostics/DiagnosticStatus.tsx
@@ -32,7 +32,7 @@ import { makeStyles } from "tss-react/mui";
 
 import Stack from "@foxglove/studio-base/components/Stack";
 import { openSiblingPlotPanel } from "@foxglove/studio-base/panels/Plot";
-import { openSiblingStateTransitionsPanel } from "@foxglove/studio-base/panels/StateTransitions";
+import { openSiblingStateTransitionsPanel } from "@foxglove/studio-base/panels/StateTransitions/openSibling";
 import { OpenSiblingPanel } from "@foxglove/studio-base/types/panels";
 
 import { DiagnosticInfo, KeyValue, DiagnosticStatusMessage, LEVELS } from "./util";
